Moved unique day-of slectors to main selectors directory and moved to use universal selectors for demographics and appointments on day-of pages. also refactored actions to match.
