+++ src/applications/check-in/day-of/pages/CheckIn/tests/CheckIn.test.unit.spec.jsx
@@ -31,6 +31,7 @@ describe('check-in', () => {
               clinicName: 'Green Team Clinic1',
             },
           ],
+          veteranData: {},
         },
       };
       store = mockStore(initState);
