+++ src/applications/check-in/day-of/pages/EmergencyContact.jsx
@@ -10,14 +10,16 @@ import { focusElement } from 'platform/utilities/ui';
 import Footer from '../../components/Footer';
 import { seeStaffMessageUpdated } from '../../actions/day-of';
 import EmergencyContactDisplay from '../../components/pages/emergencyContact/EmergencyContactDisplay';
-import { makeSelectDemographicData } from '../hooks/selectors';
+import { makeSelectVeteranData } from '../../selectors';
 
 import { URLS } from '../../utils/navigation';
 
 const EmergencyContact = props => {
   const { router } = props;
-  const selectDemographicData = useMemo(makeSelectDemographicData, []);
-  const { emergencyContact } = useSelector(selectDemographicData);
+  const selectVeteranData = useMemo(makeSelectVeteranData, []);
+  const { demographics } = useSelector(selectVeteranData);
+  const { emergencyContact } = demographics;
+
   const {
     goToNextPage,
     jumpToPage,
