+++ src/applications/check-in/day-of/pages/NextOfKin.jsx
@@ -10,14 +10,20 @@ import { focusElement } from 'platform/utilities/ui';
 import Footer from '../../components/Footer';
 import { seeStaffMessageUpdated } from '../../actions/day-of';
 import NextOfKinDisplay from '../../components/pages/nextOfKin/NextOfKinDisplay';
-import { makeSelectDemographicData } from '../hooks/selectors';
+import { makeSelectVeteranData } from '../../selectors';
 import { URLS } from '../../utils/navigation';
 
 const NextOfKin = props => {
   const { router } = props;
-  const selectDemographicData = useMemo(makeSelectDemographicData, []);
-  const { nextOfKin } = useSelector(selectDemographicData);
-  const { jumpToPage, goToNextPage, goToPreviousPage } = useFormRouting(router);
+  const selectVeteranData = useMemo(makeSelectVeteranData, []);
+  const { demographics } = useSelector(selectVeteranData);
+  const { nextOfKin1: nextOfKin } = demographics;
+  const {
+    jumpToPage,
+    goToNextPage,
+    goToPreviousPage,
+    goToErrorPage,
+  } = useFormRouting(router);
 
   const seeStaffMessage =
     'Our staff can help you update your next of kin information.';
@@ -56,7 +62,7 @@ const NextOfKin = props => {
   );
 
   if (!nextOfKin) {
-    goToNextPage(router, URLS.ERROR);
+    goToErrorPage();
     return <></>;
   } else {
     return (
