+++ src/applications/check-in/day-of/pages/SeeStaff.jsx
@@ -6,7 +6,7 @@ import Footer from '../../components/Footer';
 import { focusElement } from 'platform/utilities/ui';
 import BackButton from '../../components/BackButton';
 
-import { makeSelectSeeStaffMessage } from '../hooks/selectors';
+import { makeSelectSeeStaffMessage } from '../../selectors';
 import TravelPayReimbursementLink from '../../components/TravelPayReimbursementLink';
 
 const SeeStaff = props => {
