+++ src/applications/check-in/day-of/pages/tests/EmergencyContact.unit.spec.jsx
@@ -22,18 +22,22 @@ describe('check in', () => {
         form: {
           pages: ['first-page', 'second-page', 'third-page', 'fourth-page'],
         },
-        emergencyContact: {
-          address: {
-            street1: '445 Fine Finch Fairway',
-            street2: 'Apt 201',
-            city: 'Fairfence',
-            state: 'Florida',
-            zip: '445545',
+        veteranData: {
+          demographics: {
+            emergencyContact: {
+              address: {
+                street1: '445 Fine Finch Fairway',
+                street2: 'Apt 201',
+                city: 'Fairfence',
+                state: 'Florida',
+                zip: '445545',
+              },
+              name: 'Leslie',
+              relationship: 'Aunt',
+              phone: '5553334444',
+              workPhone: '5554445555',
+            },
           },
-          name: 'Leslie',
-          relationship: 'Aunt',
-          phone: '5553334444',
-          workPhone: '5554445555',
         },
       },
     };
@@ -64,10 +68,15 @@ describe('check in', () => {
             pages: ['first-page', 'second-page', 'third-page', 'fourth-page'],
             currentPage: 'first-page',
           },
-          emergencyContact: {
-            ...initState.checkInData.emergencyContact,
-            phone: '',
-            relationship: '',
+          veteranData: {
+            demographics: {
+              emergencyContact: {
+                ...initState.checkInData.veteranData.demographics
+                  .emergencyContact,
+                phone: '',
+                relationship: '',
+              },
+            },
           },
         },
       };
@@ -106,6 +115,9 @@ describe('check in', () => {
           form: {
             pages: ['first-page', 'second-page', 'third-page', 'fourth-page'],
             currentPage: 'first-page',
+          },
+          veteranData: {
+            demographics: {},
           },
         },
       };
