+++ src/applications/check-in/reducers/index.js
@@ -21,9 +21,7 @@ import { recordAnswerHandler, setVeteranDataHandler } from './pre-check-in';
 import {
   APPOINTMENT_WAS_CHECKED_INTO,
   RECEIVED_APPOINTMENT_DETAILS,
-  RECEIVED_EMERGENCY_CONTACT_DATA,
   RECEIVED_DEMOGRAPHICS_DATA,
-  RECEIVED_NEXT_OF_KIN_DATA,
   TRIGGER_REFRESH,
   SEE_STAFF_MESSAGE_UPDATED,
   UPDATE_DAY_OF_CHECK_IN_FORM,
@@ -32,9 +30,7 @@ import {
 import {
   appointmentWasCheckedIntoHandler,
   receivedAppointmentDetailsHandler,
-  receivedEmergencyContactDataHandler,
   receivedDemographicsDataHandler,
-  receivedNextOfKinDataHandler,
   triggerRefreshHandler,
   seeStaffMessageUpdatedHandler,
 } from './day-of';
@@ -54,9 +50,7 @@ const handler = Object.freeze({
   [SET_VETERAN_DATA]: setVeteranDataHandler,
   [APPOINTMENT_WAS_CHECKED_INTO]: appointmentWasCheckedIntoHandler,
   [RECEIVED_APPOINTMENT_DETAILS]: receivedAppointmentDetailsHandler,
-  [RECEIVED_EMERGENCY_CONTACT_DATA]: receivedEmergencyContactDataHandler,
   [RECEIVED_DEMOGRAPHICS_DATA]: receivedDemographicsDataHandler,
-  [RECEIVED_NEXT_OF_KIN_DATA]: receivedNextOfKinDataHandler,
   [TRIGGER_REFRESH]: triggerRefreshHandler,
   [SEE_STAFF_MESSAGE_UPDATED]: seeStaffMessageUpdatedHandler,
   [UPDATE_PRE_CHECK_IN_FORM]: updateFormHandler,
