+++ src/applications/check-in/selectors/selector.unit.spec.js
@@ -1,9 +1,12 @@
 import { expect } from 'chai';
+import cloneDeep from 'platform/utilities/data/cloneDeep';
 
 import {
   makeSelectCurrentContext,
   makeSelectForm,
   makeSelectVeteranData,
+  makeSelectConfirmationData,
+  makeSelectSeeStaffMessage,
 } from './index';
 
 describe('check-in', () => {
@@ -72,6 +75,62 @@ describe('check-in', () => {
               appointmentIen: 'some-appointment-ien-2',
             },
           ],
+        });
+      });
+    });
+    describe('makeSelectConfirmationData', () => {
+      const state = {
+        checkInData: {
+          appointments: [
+            {
+              clinicPhone: '555-867-5309',
+              startTime: '2021-07-19T13:56:31',
+              facilityName: 'Acme VA',
+              clinicName: 'Green Team Clinic1',
+            },
+          ],
+          context: {
+            appointment: {
+              appointmentIen: 'some-ien',
+            },
+            token: 'foo',
+          },
+        },
+      };
+      it('returns appointment confirmation data', () => {
+        const selectConfirmationData = makeSelectConfirmationData();
+        expect(selectConfirmationData(state)).to.eql({
+          appointments: [
+            {
+              clinicName: 'Green Team Clinic1',
+              clinicPhone: '555-867-5309',
+              facilityName: 'Acme VA',
+              startTime: '2021-07-19T13:56:31',
+            },
+          ],
+          selectedAppointment: {
+            appointmentIen: 'some-ien',
+          },
+        });
+      });
+      it('returns empty when appointment data is not available', () => {
+        const partialState = cloneDeep(state);
+        delete partialState.checkInData.appointments;
+        delete partialState.checkInData.context.appointment;
+        const selectConfirmationData = makeSelectConfirmationData();
+        expect(selectConfirmationData(partialState)).to.eql({});
+      });
+    });
+    describe('makeSelectSeeStaffMessage', () => {
+      const state = {
+        checkInData: {
+          seeStaffMessage: 'Test message',
+        },
+      };
+      it('returns see staff message', () => {
+        const selectSeeStaffMessage = makeSelectSeeStaffMessage();
+        expect(selectSeeStaffMessage(state)).to.eql({
+          message: 'Test message',
         });
       });
     });
